new simple8b

+++ simple8b/tsz_data.go
@@ -2,7 +2,7 @@ package simple8b
 
 type Point struct {
 	V float64
-	T uint32
+	T uint64
 }
 
 // 120 points every 60s
